// part9/courseinfo-typescript/src/components/Part.tsx
import type { CoursePart } from '../types';

interface PartProps {
    part: CoursePart;
}

const assertNever = (value: never): never => {
    throw new Error(
        `Unhandled discriminated union member: ${JSON.stringify(value)}`
    );
};

const Part = ({ part }: PartProps) => {

    switch (part.kind) {
    case 'basic':
        return (
            <div className='part'>
                <h3>{part.name} {part.exerciseCount}</h3>
                <p><i>{part.description}</i></p>
            </div>
        );
    case 'group':
        return (
            <div className='part'>
                <h3>{part.name} {part.exerciseCount}</h3>
                <p>project exercises {part.groupProjectCount}</p>
            </div>
        );
    case 'background':
        return (
            <div className='part'>
                <h3>{part.name} {part.exerciseCount}</h3>
                <p><i>{part.description}</i></p>
                <p>submit to {part.backgroundMaterial}</p>
            </div>
        );
    case 'special':
        return (
            <div className='part'>
                <h3>{part.name} {part.exerciseCount}</h3>
                <p><i>{part.description}</i></p>
                <p>required skills: {part.requirements.join(', ')}</p>
            </div>
        );
    default:
        return assertNever(part);
    }
};

export default Part;